+++ modules/cache-common/src/test/java/org/opensearch/cache/common/tier/TieredSpilloverCachePluginTests.java
@@ -9,18 +9,15 @@
 package org.opensearch.cache.common.tier;
 
 import org.opensearch.common.cache.ICache;
-import org.opensearch.common.cache.provider.CacheProvider;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.util.Map;
-
-import static org.mockito.Mockito.mock;
 
 public class TieredSpilloverCachePluginTests extends OpenSearchTestCase {
 
     public void testGetCacheFactoryMap() {
         TieredSpilloverCachePlugin tieredSpilloverCachePlugin = new TieredSpilloverCachePlugin();
-        Map<String, ICache.Factory> map = tieredSpilloverCachePlugin.getCacheFactoryMap(mock(CacheProvider.class));
+        Map<String, ICache.Factory> map = tieredSpilloverCachePlugin.getCacheFactoryMap();
         assertNotNull(map.get(TieredSpilloverCache.TieredSpilloverCacheFactory.TIERED_SPILLOVER_CACHE_NAME));
         assertEquals(TieredSpilloverCachePlugin.TIERED_CACHE_SPILLOVER_PLUGIN_NAME, tieredSpilloverCachePlugin.getName());
     }